Smaller balloons based on a new Figma export.
Force changes:
- weaker repulsion `strength` from `forceManyBody` for tighter spacing
- smaller collision `radius` (tested w/positive fMB.strength so it should work once balloon groups start interacting)
- tweaked `forceCenter`

Also extracted the Balloon rendering chain into an exported function in the new `Balloon.svelte`. Note the use of `context="module"` in that file. This script prop allows us to import the function in App.svelte's `<script>` tag.

This commit comes after hours of trying to get Svelte and d3 to behave nicely together. In particular, I was hoping to have the balloons be static HTML elements, imported by Svelte, and tagged with the appropriate node id's form the source data, after which point, hypothetically, d3 should take over and place them where epxected. Instead, all I get is elements that either won't render or won't reposition. 😩

Why did I want to do that? Because it's tedious work, converting an exported SVG into a d3 function chain! Busy work. But that's ok. It's not worth gating the project to fix, I don't think, since it should be pretty easy to swap out if I ever find a solution.