Figured out how to effectively split concerns between _svelte_ and _d3_!
# The Goal

I wanted to use Figma's exported svg in its original XML, rather than having to try to rebuild it all in a horrific .attr() chain.

The main benefit is reduced effort as I can copy/paste the source SVG into Balloon.svelte file and, with minor tweaks to integrate the single (for now) component prop (`id`), it's off to the races!

Another benefit is that I can let Svelte remain in charge of creating and destroying DOM nodes, and d3 can just be in charge of the physics.