🧹 Extract functions => `applyBasicForces` & `applyHeightForce`
The goal here is to get closer to being able to start with the y force positioning everything up top off-screen, then two swap the forces so that the balloons drop down from offstage center. Right now, they always start at 0,0, which is annoying…mainly because it's not what I originally intended. (It actually looks pretty cool…but I want to learn how to wield d3 more intentionally, so…let's do this!)