Start with very low alphaDecay so that the sim doesn't end so far from the "final" balloon destinations.
The motivator here is that the default behavior causes the balloons to freeze before they get to their destination. For "perfect" placement, but a more hurried entry, change it to `alphaDecay(0)`, which will get the balloons to their near-final, "fully accurate" positions lickety split.

There's a tradeoff here. With the default decay, balloons that came in a little later, thanks to staggered entry, stop moving further to the left than the first balloons in. This gives the whole group a slight slant, like a `/`, and causes them to trend too far left. On the other hand, a super low- or a zero decay results in the correct positions, but too-fast entry. About 1.5 seconds' worth of .007 decay — followed by the default 0.228 — seems to offer a happy medium.