Removed timeout in `dropper`
Not sure why I had it in there in the first place — maybe so the balloons wouldn't be moving so fast on entry? At any rate, when I first played around with a "+" button, the full second delay was causing newly-added balloons to enter way too slowly such that they never made it to the center.