Grid off
Now that I'm doing less wonky stuff with the sim forces and the balloons aren't landing so far off center, I'm giving this a rest. Later, I'll need to revisit to figure out what's going on with the y values…i.e., why, when there are only 5 balloons (1 at each of the levels) are they not vertically-centered within each level?