Factored out position, scaling, & flipping from CloudLayer & Grass
Note the lack of a 🚧 icon in the main commit line. That's because this is definitely not a pure refactor…though it is *close*. Afaict, the only functional difference occurs when performing a horizontal resize on the window (dragging the side of the window, as opposed to the corner or the bottom).

Previously, horizontal resizing would leave clouds currently-visible "in place". Now, since any window resize causes the entire "Repeater" to react, the clouds get new vertical positions as you drag the side of the window in and out. This results in a sort of jittery motion, similary to what you'd see when resizing from the corner or bottom previously. Which is fine.