🚧 Factored out grassHeight and cloudLayerHeight
Why? Because I'm about to try implementing "collisions" between the balloons and the clouds/grass, but faked such that it's really just a matter of whether or not the balloons cross into the rect in which the clouds/grass live. If that doesn't make sense, well…skip to the next commit. ;)