Experiment: "Form Editor"
At this point, it's got nothing to do with the actual form editor, and everything to do with attempting to place an SVG balloon in a responsive manner. Got it more or less working using Svelte to set the values. Next up: see if I can get the same behaviors out of CSS & `calc()`!